2020/Chatbot_Fall2020/Cloud Functions/webhook/requirements.txt
# Function dependencies, for example:
# package>=version
line-bot-sdk==1.16.0
dialogflow==1.0.0
